Attachments, Coordinate
Attachments
New type and module for structuring how Edge attachments can be set and edited.
It has Unset variants so editing an attachment can be reflected in the data, this also means that encoding is now a Result, as an Unset attachment can't be encoded.

Coordinate
A WIP system for encapsulating the coordinate system used by Obsidian canvases.

Edge
- From/to/node/side fields and functions have been renamed to attachment as they are now encapsulated in the Attachments module.
- 'change' terminology changed to 'set' to read better.

Node
- 'change' terminology changed to 'set' to read better.